Added Character.java, Class.java, Race.java, Item.java & Preset.java
Character.java contains the contents of Race.java, and will soon
contain the contents of Class.java. This should make both Race & Class
redundant, but I'm keeping it in this commit for legacy sakes.

Currently everything is a mess, but I won't have one of you looking like
you developed everything because I only commit once I've finished what
I'm working on (save for this specific case, and potentially more in
the future).