Added a couple of player presets for testing purposes- if you want to use the character creator, just uncomment the 6 lines before the game loop
Added characters, an enemy array, an all entity array (for things like
SELECT)

Made it so that you can't equip items for enemies, but you can show
their inventory, look at their stats, etc

Made it so that presets can now be for players or entities. When making
a preset, make sure you put in the player flag.

Added new command:
START -> starts the preset combat scenario (combat hasn't been added
yet, so it just loads them)

Added a loadTestScenario function which is called when START is used
at the moment. It just builds all the preset enemies and throws them
in the enemy and allEntities arrays