Added the basic combat system:
Character class now has bool alive which tracks if the player is dead

The gameplay loops to check for dead players and removes them from the
game if they are.

death() function in main displays a death message for dead characters

Character.attack() added, dice rolls to attack enemies and hits them if
successful. Otherwise it just displays a missed message.

Character.calculateDamage() added damage calculations based on strength
or wisdom depending on if the weapon is ranged or melee

Couple of other extra goodies that have escaped me